Generic Programming
Using Dynamic memory Allocation and Generic Programming
we build a Hash Tables data structures that support these three operations:

1. Adding an element
2. Removing an element
3. Searching for an element

the size of the array is dynamic, i.e. the size of the array can grow.
Mapping of an object to the appropriate location in this array is accomplished using a hash function:
i=d*H(k,m)